requirements.txt
asgiref==3.3.1
attrs==20.3.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
dateparser==1.0.0
Django==3.1.7
django-mysql==3.11.1
django-uuid-upload-path==1.0.0
djangorestframework==3.12.2
docx2txt==0.8
docxtpl==0.11.3
en-core-web-sm==2.2.5
idna==2.10
imgkit==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
lxml==4.6.3
MarkupSafe==1.1.1
murmurhash==1.0.5
nltk==3.5
numpy==1.20.1
pandas==1.2.3
pdfkit==0.6.1
pdfminer==20191125
pdfminer.six==20200517
pdfplumber==0.5.27
pexpect==4.2.1
phonenumbers==8.12.19
Pillow==8.1.2
plac==1.1.3
preshed==3.0.5
ptyprocess==0.7.0
pycryptodome==3.10.1
PyPDF2==1.26.0
pyresparser==1.0.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
regex==2021.3.17
requests==2.25.1
resume-parser==0.8
six==1.15.0
sortedcontainers==2.3.0
spacy==2.2.4
spacy-legacy==3.0.1
sqlparse==0.4.1
srsly==1.0.5
stemming==1.0.1
thinc==7.4.0
tika==1.24
tqdm==4.59.0
tzlocal==2.1
urllib3==1.26.4
Wand==0.6.6
wasabi==0.8.2
wkhtmltopdf==0.2
